Add device list

+++ notify.py
@@ -3,6 +3,8 @@
 import ssl
 import json
 import socket
+import binascii
+from pathlib import Path
 from OpenSSL import crypto
 from argparse import ArgumentParser
 
@@ -16,17 +18,22 @@ parser.add_argument("-c", "--config", dest="config", default=None, help="custom 
 # Parse args and delegate
 namespace = parser.parse_args()
 config = get_config(namespace.config)
-p12 = load_certificates(config)
-with open("files/certificates.pem", "wb") as file:
-    file.write(crypto.dump_certificate(crypto.FILETYPE_PEM, p12.get_certificate()))
-    file.write(crypto.dump_privatekey(crypto.FILETYPE_PEM, p12.get_privatekey()))
+certificates = Path(config["files"]["certificatesPath"].rstrip("p12") + "pem")
 
+# Check if the certificate needs to be converted
+if not certificates.exists():
+    p12 = load_certificates(config)
+    with open("files/certificates.pem", "wb") as file:
+        file.write(crypto.dump_certificate(crypto.FILETYPE_PEM, p12.get_certificate()))
+        file.write(crypto.dump_privatekey(crypto.FILETYPE_PEM, p12.get_privatekey()))
+
+# Get input for the payload
 title = input("Title: ")
 body = input("Body: ")
 args = map(lambda x: x.strip(), input("Slug: ").split(","))
 
-message = json.dumps(
-{
+# Generate the payload body
+message = json.dumps({
     "aps": {
         "alert": {
             "title": title,
@@ -37,12 +44,12 @@ message = json.dumps(
     },
 })
 
+# Pack the message
 
-print(message)
 
+#
 hostname = "ssl://gateway.push.apple.com:2195"
 context = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
-# context.load_verify_locations(cafile="files/intermediate.pem")
 context.load_cert_chain("files/certificates.pem")
 
 with socket.socket(socket.AF_INET, socket.SOCK_STREAM, 0) as sock:
